ASoC: soc-core: remove snd_soc_rtdcom_list
Current ALSA SoC is using modern style dai_link,
at that time, it is indicating number of CPU/Codec/Platform components.
These components are connected to rtd.
This means, it is possible to know how many components will be connected
to rtd when it was allocated,
and we can avoid to use link list and use component pointer array instead.

Signed-off-by: Kuninori Morimoto <kuninori.morimoto.gx@renesas.com>